MAINT: Drop BC R101 support
