+++ src/settings/settings_screen.ts
@@ -2,7 +2,7 @@ import { MBS_MOD_API, waitFor, logger } from "../common";
 import { bcLoaded } from "../common_bc";
 import { MBSScreen, ScreenProxy } from "../screen_abc";
 import { FWSelectScreen, loadFortuneWheelObjects } from "../fortune_wheel";
-import { NewItemsScreen, NEW_ASSETS_VERSION, NEW_ASSETS } from "../new_items_screen";
+import { NEW_ASSETS_VERSION, NEW_ASSETS } from "../new_items_screen";
 
 import {
     pushMBSSettings,
@@ -153,22 +153,16 @@ export class MBSPreferenceScreen extends MBSScreen {
                     DrawText(`MBS: Show new R${NEW_ASSETS_VERSION} items`, x + descriptionOffset, y + h / 2, "Black");
                 },
                 click: () => {
-                    if (typeof Shop2 === "undefined") {
-                        const subScreen = new NewItemsScreen(this);
-                        this.children.set(subScreen.screen, subScreen);
-                        subScreen.load();
-                    } else {
-                        const background = ServerPlayerIsInChatRoom() ? ChatRoomData?.Background : undefined;
-                        const prevScreen: [ModuleType, string] = ServerPlayerIsInChatRoom() ? ["Online", "ChatRoom"] : ["Character", "Preference"];
-                        this.exitScreens(true);
-                        Shop2Vars.Mode = "Preview";
-                        Shop2Vars.Filters.MBS_VersionFilter = (item) => NEW_ASSETS[`${item.Asset.Group.Name}${item.Asset.Name}`] ? ["Buy", "Sell", "Preview"] : [];
-                        Shop2.Init(background, prevScreen);
-                        ServerBeep = {
-                            Message: "The MBS \"Show new items\" button will be removed in R103; use the Club Shop or /shop chat command instead",
-                            Timer: CommonTime() + 10000,
-                        };
-                    }
+                    const background = ServerPlayerIsInChatRoom() ? ChatRoomData?.Background : undefined;
+                    const prevScreen: [ModuleType, string] = ServerPlayerIsInChatRoom() ? ["Online", "ChatRoom"] : ["Character", "Preference"];
+                    this.exitScreens(true);
+                    Shop2Vars.Mode = "Preview";
+                    Shop2Vars.Filters.MBS_VersionFilter = (item) => NEW_ASSETS[`${item.Asset.Group.Name}${item.Asset.Name}`] ? ["Buy", "Sell", "Preview"] : [];
+                    Shop2.Init(background, prevScreen);
+                    ServerBeep = {
+                        Message: "The MBS \"Show new items\" button will be removed in R103; use the Club Shop or /shop chat command instead",
+                        Timer: CommonTime() + 10000,
+                    };
                 },
             },
             wheelConfig: {
